Add support for custom buffers
Allow users of the API to provide the message buffers to work with. This
allows IO to be performed directly on those buffers, skipping a copy
operation.

Since the server uses the write_data function to write keydata,
introduce a flush_data function that performs this task instead, which
allows the custom buffers to receive this data as well.

If no buffer is provided by the user, use a vector as before, but
wrapped so the ksnp_buffer interface can be used. The use of dequeue is
replaced by vector as well.

Change-Id: I803180b0d5c7e7a0d5c910128ed09adbb8861df3